create admin-controller, service, route and doc
