+++ src/controllers/organization.controller.ts
@@ -33,42 +33,6 @@ const registerOrganization = async (req: Request, res: Response, next: NextFunct
   });
 };
 
-export const getAllOrganizations = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  const user = req.user!;
-
-  if (user.siteRole !== 'ADMIN') {
-    logger.warn('⚠️ Unauthorized access attempt to admin organizations list', {
-      userId: user.id,
-      role: user.siteRole,
-    });
-    return next(httpError(403, 'Forbidden: Admin access required', ErrorCode.FORBIDDEN));
-  }
-
-  const page = parseInt(req.query.page as string, 10) || 1;
-  const limit = parseInt(req.query.limit as string, 10) || 10;
-  const search = (req.query.search as string)?.trim() || '';
-
-  logger.info('👤 Admin fetching all organizations list', {
-    adminId: user.id,
-    page,
-    limit,
-    ...(search && { search }),
-  });
-
-  const result = await organizationServices.getAllOrganizationsForAdmin(page, limit, search);
-
-  return res.status(200).json({
-    status: 'success',
-    code: SuccessCode.ORGANIZATION_DATA_RETRIEVED,
-    message: 'Organizations retrieved successfully',
-    ...result,
-  });
-};
-
 const getOrganizationById = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
 
@@ -440,7 +404,6 @@ const getJoinRequestById = async (req: Request, res: Response, next: NextFunctio
 
 export const organizationControllers = {
   registerOrganization: asyncHandler(registerOrganization),
-  getAllOrganizations: asyncHandler(getAllOrganizations),
   getOrganizationById: asyncHandler(getOrganizationById),
   getOrganizationsByName: asyncHandler(getOrganizationsByName),
   inviteMemberToOrganization: asyncHandler(inviteMemberToOrganization),
