+++ src/routes/api/index.ts
@@ -9,4 +9,5 @@ export { healthRoute } from './health.route';
 export { donateRoute } from './donate.route';
 export { organizationRoute } from './organization.route';
 export { supportRoute } from './support.route';
+export { adminRoute } from './admin.route';
