+++ src/routes/api/organization.route.ts
@@ -12,12 +12,6 @@ organizationRoute.post('/create',
   organizationControllers.registerOrganization
 );
 
-organizationRoute.get(
-  '/admin/organizations',
-  authenticateUser,
-  organizationControllers.getAllOrganizations 
-);
-
 organizationRoute.get('/', organizationControllers.getOrganizationsByName);
 
 organizationRoute.get('/:id', organizationControllers.getOrganizationById);
